% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.query.R
\name{repo.query}
\alias{repo.query}
\title{Retrieve SPARQL results from grlc repo}
\usage{
repo.query(query, number)
}
\arguments{
\item{object}{from repo summary; object number}
}
\description{
This function retrieves data from the grlc-repo sparql query
}
\examples{
repo.query()

}
\keyword{grlc}
