% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.summary.R
\name{repo.summary}
\alias{repo.summary}
\title{Retrieve SPARQL queries from grlc repo}
\usage{
repo.summary(git)
}
\arguments{
\item{a}{grlc repo owner/project}
}
\description{
This function retrieves the available sparql queries from a grlc repo
}
\examples{
repo.summary()

}
\keyword{grlc}
