% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_query.R
\name{repo_query}
\alias{repo_query}
\title{Retrieve SPARQL results from grlc repo}
\usage{
repo_query(query, number)
}
\arguments{
\item{object}{from repo summary; object number}
}
\description{
This function retrieves data from the grlc-repo sparql query
}
\examples{
repo_query()

}
\keyword{grlc}
