% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uri_value.R
\name{uri_value}
\alias{uri_value}
\title{Retrieve the 'value' (last part) of a uri}
\usage{
uri_value(x)
}
\arguments{
\item{a}{uri}
}
\value{
The value after the last '/' of a \uri_value{x}
}
\description{
Retrieve the 'value' (last part) of a uri
}
\examples{
uri_value("http://example.com/myvalues/1234")
uri_value("http://example.com/myvalues/bookstore/") # warns for trailing slash

}
